functions cleanup moved to lib
